It's coming together!
Starting stuff with map visibility! Really coming close.
Change so that collide in check doesn't show every time, I don't need
that anymore. Also work on math for the visibility on the map. It's
almost there. Also the test placeholder and test slime should be
initailized at the same tile.