fixy fixy
Map display overhaul. Not using the TMX stuff anymore. Key usage added
but not polished, also fixed some issues with the collision map not
transferring because objects are really reference objects, so issues
with idmap = collisionmap; But they're all fixed. Also fixed the part
where the door didn't trigger, and then you don't collide with the door
making a null pointer with the collision map, but now it collides and
goes to the next floor. The display seems much more smooth without the
TMX! Woot!